config: break out _util.DomainGlobSet; use it for federation_domain_whitelist
